Fix merge markers, clean imports, and dedupe VehicleData
