+++ apps/ain-valuation-engine/src/ain-backend/whisperService.ts
@@ -1,5 +1,5 @@
 import OpenAI from 'openai';
-import logger from '../utils/logger.js';
+import logger from '../utils/logger';
 import fs from 'fs';
 import path from 'path';
