+++ apps/ain-valuation-engine/src/services/carApiService.ts
@@ -1,5 +1,5 @@
-import { NormalizedVehicle } from "../utils/normalizeVehicleData.js";
-import logger from "../utils/logger.js";
+import { NormalizedVehicle } from "../utils/normalizeVehicleData";
+import logger from "../utils/logger";
 
 export async function carApiService(vehicle: NormalizedVehicle): Promise<{ [key: string]: any }> {
   logger.info("Calling Car API Service", { vin: vehicle.vin });
